Cleaning `opts.lua`
Reformatted it to very visibly separate globals and options. I also
replaced all instances of `vim.o` with `vim.opt` as it's supposed to be
a "modern" replacement and is recommended to use. I'll use `vim.opt`
whenever possible from now on. I'm not entirely sure how it's better but
it is, if you wanna know go Google or smth.